tools/MetlifeContractCapture/requirements.txt
# 화면 캡처
mss>=9.0.0
pyautogui>=0.9.54
pillow>=10.0.0

# 중복 감지
imagehash>=4.3.0

# AI Vision / OCR
anthropic>=0.45.0
httpx>=0.27.0

# 데이터 처리
pandas>=2.0.0
openpyxl>=3.1.0

# CLI
click>=8.1.0
rich>=13.0.0

# 설정
pyyaml>=6.0.0
